settings: properly update all devices attached to an email
Mongoose really sucks and when you use findByIdAndUpdate inside
of a for loop, mongoose decides to die, because why not?

I swear to gosh this ODM library will be the end of me.

anyways, using this cool new function called Device.update
that allows us to update many models at once. Cool

Signed-off-by: Sun <sunrunr@hwsrv-644685.hostwindsdns.com>